Init commit
Info box is waited on until present, and then prepended into
The info box seems to be the best fit. Anything pertaining to the video container
really does not seem to want anything messing with it. The info box
adheres to the same size of the video, and is directly below it

The oscilloscope functions as just a bunch of line segments

The button functions, but the oscilloscope just changes color
immediately